+++ app/src/main/java/in/surajsau/jisho/data/FileProvider.kt
@@ -1,8 +1,10 @@
 package `in`.surajsau.jisho.data
 
+import `in`.surajsau.jisho.base.rotate
 import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
+import android.media.ExifInterface
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import java.io.File
@@ -15,9 +17,17 @@ class FileProviderImpl @Inject constructor(private val context: Context): FilePr
         val cacheDir = context.externalCacheDir ?: error("No external cache found")
         val bitmap = runCatching {
             val imageFile = File(cacheDir, fileName)
+            val exif = ExifInterface(imageFile.absolutePath)
             val inputStream = FileInputStream(imageFile)
-            BitmapFactory.decodeStream(inputStream)
-        }
+            val fileBitmap = BitmapFactory.decodeStream(inputStream)
+
+            return@runCatching when (exif.getAttributeInt(ExifInterface.TAG_ORIENTATION, ExifInterface.ORIENTATION_NORMAL)) {
+                ExifInterface.ORIENTATION_ROTATE_90 -> fileBitmap.rotate(90)
+                ExifInterface.ORIENTATION_ROTATE_180 -> fileBitmap.rotate(180)
+                ExifInterface.ORIENTATION_ROTATE_270 -> fileBitmap.rotate(270)
+                else -> fileBitmap
+            }
+        }.onFailure { it.printStackTrace() }
 
         emit(bitmap.getOrElse { error(it) })
     }
