+++ app/src/main/java/in/surajsau/jisho/ui/styletransfer/StyleTransferScreen.kt
@@ -1,37 +1,15 @@
 package `in`.surajsau.jisho.ui.styletransfer
 
-import `in`.surajsau.jisho.base.LocalBitmapCache
 import `in`.surajsau.jisho.base.use
 import `in`.surajsau.jisho.ui.base.AskPermissionScreen
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.clipToBounds
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalLifecycleOwner
-import androidx.compose.ui.unit.dp
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
 
 @Composable
 fun StyleTransferScreen(
@@ -57,7 +35,10 @@ fun StyleTransferScreen(
         }
     }
 
-    Box(modifier = modifier.fillMaxSize()) {
+    Box(modifier = modifier
+        .fillMaxSize()
+        .background(Color.White)
+    ) {
 
         AskPermissionScreen(
             modifier = Modifier.fillMaxSize(),
@@ -79,14 +60,19 @@ fun StyleTransferScreen(
                     StylePreviewScreen(
                         modifier = Modifier.fillMaxSize(),
                         imagePreview = {
-                            ImagePreview(bitmap = screenMode.image, showLoader = screenMode.showLoading)
+                            ImagePreview(
+                                bitmap = screenMode.image,
+                                showLoader = screenMode.showLoading,
+                                modifier = Modifier.fillMaxSize()
+                            )
                         },
                         previewGallery = {
                             PreviewGallery(
                                 previews = screenMode.stylePreviews,
                                 onStyleSelected = { fileName ->
                                     event(StyleTransferViewModel.Event.StyleSelected(fileName))
-                                }
+                                },
+                                modifier = Modifier.fillMaxSize()
                             )
                         }
                     )
@@ -96,118 +82,4 @@ fun StyleTransferScreen(
         }
     }
 
-}
-
-@Composable
-fun StylePreviewScreen(
-    imagePreview: @Composable () -> Unit,
-    previewGallery: @Composable () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-
-    Column(modifier = modifier.background(Color.Black)) {
-
-        Box(modifier = Modifier
-            .fillMaxWidth()
-            .weight(1f)
-        ) {
-            imagePreview()
-        }
-
-        previewGallery()
-    }
-
-}
-
-@Composable
-fun ImagePreview(
-    bitmap: Bitmap?,
-    modifier: Modifier = Modifier,
-    showLoader: Boolean = false,
-) {
-
-    Box(modifier = modifier) {
-        if (bitmap != null) {
-            Image(
-                bitmap = bitmap.asImageBitmap(),
-                contentDescription = null,
-                modifier = Modifier.fillMaxSize()
-            )
-        }
-
-        if (showLoader) {
-            CircularProgressIndicator(modifier = Modifier.align(Alignment.Center))
-        }
-    }
-}
-
-@Composable
-fun PreviewGallery(
-    previews: List<String>,
-    modifier: Modifier = Modifier,
-    onStyleSelected: (String) -> Unit
-) {
-
-    LazyRow(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(16.dp)
-    ) {
-        itemsIndexed(previews) { index, fileName ->
-            PreviewGalleryItem(
-                fileName = fileName,
-                onClick = { onStyleSelected.invoke(it) },
-                modifier = Modifier
-                    .width(96.dp)
-                    .height(96.dp)
-                    .padding(
-                        start = if (index == 0) 16.dp else 0.dp,
-                        end = 16.dp
-                    )
-            )
-        }
-    }
-}
-
-@Composable
-fun PreviewGalleryItem(
-    fileName: String,
-    onClick: (String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-
-    val bitmapCache = LocalBitmapCache.current
-
-    val context = LocalContext.current
-
-    var previewBitmap by remember { mutableStateOf<Bitmap?>(null) }
-
-    LaunchedEffect(fileName) {
-        if (!bitmapCache.has(fileName)) {
-            launch(Dispatchers.IO) {
-                val bitmapStream = context.assets.open(fileName)
-                val bitmap = BitmapFactory.decodeStream(bitmapStream)
-
-                bitmapCache.save(fileName, bitmap)
-                bitmapStream.close()
-
-                previewBitmap = bitmap
-            }
-        } else {
-            previewBitmap = bitmapCache.get(fileName)
-        }
-    }
-
-    if (previewBitmap != null) {
-        Image(
-            bitmap = previewBitmap!!.asImageBitmap(),
-            contentDescription = null,
-            modifier = modifier
-                .border(width = 2.dp, color = Color.White, shape = RoundedCornerShape(4.dp))
-                .clip(RoundedCornerShape(4.dp))
-                .clipToBounds()
-                .clickable { onClick.invoke(fileName) },
-            contentScale = ContentScale.Crop
-        )
-    }
 }
